Moves game config interfaces to interfaces file
