+++ src/app/game/services/game-config.service.ts
@@ -1,16 +1,6 @@
+import { GameLevelConfig, ImageScoreConfig } from '@/app/shared/models/interfaces';
 import { Injectable } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
-
-export interface GameLevelConfig {
-  difficultyId: number;
-  rounds: number;
-  secondsPerRound: number;
-  timeToStartClassify: number;
-}
-export interface ImageScoreConfig {
-  maxScore: number;
-  scoreDecrement: number;
-}
 
 @Injectable({
   providedIn: 'root',
